fix: contact Joi schemas path changed
